feat: Add module about calculator
